+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/Build.java
@@ -10,6 +10,8 @@ import java.io.IOException;
 
 import org.apache.http.client.HttpResponseException;
 
+import com.google.common.collect.ImmutableMap;
+
 public class Build extends BaseModel {
 
     /**